# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Estimation Engine",
    "description": "Construction cost estimation engine API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/calculations": {
      "post": {
        "tags": [
          "calculations"
        ],
        "summary": "Create Calculation",
        "description": "Create a new calculation job.",
        "operationId": "create_calculation_v1_calculations_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalculationRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calculations/{calculation_id}": {
      "get": {
        "tags": [
          "calculations"
        ],
        "summary": "Get Calculation",
        "description": "Get full calculation result or progress.",
        "operationId": "get_calculation_v1_calculations__calculation_id__get",
        "parameters": [
          {
            "name": "calculation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calculations/{calculation_id}/status": {
      "get": {
        "tags": [
          "calculations"
        ],
        "summary": "Get Calculation Status",
        "description": "Lightweight polling endpoint for calculation status.",
        "operationId": "get_calculation_status_v1_calculations__calculation_id__status_get",
        "parameters": [
          {
            "name": "calculation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/calculations/{calculation_id}/cancel": {
      "post": {
        "tags": [
          "calculations"
        ],
        "summary": "Cancel Calculation",
        "description": "Request cancellation of a calculation job.",
        "operationId": "cancel_calculation_v1_calculations__calculation_id__cancel_post",
        "parameters": [
          {
            "name": "calculation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Calculation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalculationJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CalculationJobResponse": {
        "properties": {
          "calculation_id": {
            "type": "string",
            "title": "Calculation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent",
            "default": 0
          },
          "processed_items": {
            "type": "integer",
            "title": "Processed Items",
            "default": 0
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items",
            "default": 0
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "failed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          }
        },
        "type": "object",
        "required": [
          "calculation_id",
          "status"
        ],
        "title": "CalculationJobResponse"
      },
      "CalculationRequest": {
        "properties": {
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "region": {
            "$ref": "#/components/schemas/RegionInput"
          },
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SpecificationItemInput"
            },
            "type": "array",
            "minItems": 1,
            "title": "Items"
          },
          "request_meta": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Meta"
          }
        },
        "type": "object",
        "required": [
          "region",
          "currency",
          "items"
        ],
        "title": "CalculationRequest",
        "example": {
          "currency": "RUB",
          "items": [
            {
              "category": "masonry",
              "code": "aerated_concrete_block_300_d500",
              "id": "item-1",
              "kind": "material",
              "name": "Газобетонный блок D500 300мм",
              "quantity": 9,
              "unit": "m3"
            },
            {
              "category": "masonry",
              "code": "block_masonry",
              "id": "item-2",
              "kind": "work",
              "name": "Кладка газобетонных блоков",
              "quantity": 30,
              "unit": "m2"
            }
          ],
          "region": {
            "city": "Moscow",
            "country_code": "RU",
            "region_code": "RU-MOW"
          }
        }
      },
      "CalculationResultResponse": {
        "properties": {
          "calculation_id": {
            "type": "string",
            "title": "Calculation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SummaryOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "items": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/PricedItemOutput"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items"
          },
          "region": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "assumptions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assumptions"
          },
          "diagnostics": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostics"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent",
            "default": 0
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "failed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failed At"
          }
        },
        "type": "object",
        "required": [
          "calculation_id",
          "status"
        ],
        "title": "CalculationResultResponse"
      },
      "CalculationStatusResponse": {
        "properties": {
          "calculation_id": {
            "type": "string",
            "title": "Calculation Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent",
            "default": 0
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          }
        },
        "type": "object",
        "required": [
          "calculation_id",
          "status"
        ],
        "title": "CalculationStatusResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PricedItemOutput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "material",
              "work"
            ],
            "title": "Kind"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "title": "Unit"
          },
          "pricing": {
            "$ref": "#/components/schemas/PricingInfoOutput"
          },
          "totals": {
            "$ref": "#/components/schemas/TotalsOutput"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "code",
          "name",
          "quantity",
          "unit",
          "pricing",
          "totals"
        ],
        "title": "PricedItemOutput"
      },
      "PricingInfoOutput": {
        "properties": {
          "average_unit_price": {
            "type": "number",
            "title": "Average Unit Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "price_unit": {
            "type": "string",
            "title": "Price Unit"
          },
          "sources_count": {
            "type": "integer",
            "title": "Sources Count"
          },
          "min_unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Unit Price"
          },
          "max_unit_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Unit Price"
          },
          "pricing_method": {
            "type": "string",
            "enum": [
              "exact_match",
              "country_fallback",
              "unit_conversion",
              "category_fallback",
              "coefficient_fallback",
              "unpriced",
              "requires_manual_review"
            ],
            "title": "Pricing Method"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "high",
              "medium",
              "low",
              "none"
            ],
            "title": "Confidence"
          },
          "match_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Path"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "unit_converted": {
            "type": "boolean",
            "title": "Unit Converted",
            "default": false
          },
          "original_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Unit"
          },
          "resolution_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Level"
          },
          "sources_queried": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources Queried"
          }
        },
        "type": "object",
        "required": [
          "average_unit_price",
          "currency",
          "price_unit",
          "sources_count",
          "pricing_method",
          "confidence"
        ],
        "title": "PricingInfoOutput"
      },
      "RegionInput": {
        "properties": {
          "country_code": {
            "type": "string",
            "maxLength": 2,
            "minLength": 2,
            "title": "Country Code"
          },
          "region_code": {
            "type": "string",
            "minLength": 1,
            "title": "Region Code"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          }
        },
        "type": "object",
        "required": [
          "country_code",
          "region_code"
        ],
        "title": "RegionInput"
      },
      "SpecificationItemInput": {
        "properties": {
          "id": {
            "type": "string",
            "minLength": 1,
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "material",
              "work"
            ],
            "title": "Kind"
          },
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "quantity": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Quantity"
          },
          "unit": {
            "type": "string",
            "minLength": 1,
            "title": "Unit"
          },
          "category": {
            "type": "string",
            "minLength": 1,
            "title": "Category"
          },
          "metadata": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "code",
          "name",
          "quantity",
          "unit",
          "category"
        ],
        "title": "SpecificationItemInput"
      },
      "SummaryOutput": {
        "properties": {
          "grand_total": {
            "type": "number",
            "title": "Grand Total"
          },
          "materials_total": {
            "type": "number",
            "title": "Materials Total"
          },
          "works_total": {
            "type": "number",
            "title": "Works Total"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "priced_items": {
            "type": "integer",
            "title": "Priced Items"
          },
          "fallback_items": {
            "type": "integer",
            "title": "Fallback Items"
          },
          "unpriced_items": {
            "type": "integer",
            "title": "Unpriced Items"
          }
        },
        "type": "object",
        "required": [
          "grand_total",
          "materials_total",
          "works_total",
          "currency",
          "total_items",
          "priced_items",
          "fallback_items",
          "unpriced_items"
        ],
        "title": "SummaryOutput"
      },
      "TotalsOutput": {
        "properties": {
          "line_total": {
            "type": "number",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "line_total"
        ],
        "title": "TotalsOutput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}